samples/python/02.echo-bot/requirements.txt
botbuilder-integration-aiohttp>=4.14.0
python-dotenv==1.0.1